Update remote grid based on diff from previous
Previously we were updating the entire grid based on the remote state.
This causes a race such that if both users update different squares at the
same time (or even if they are moving quickly around the grid) their edits may
not show up.

This does not fix the problem of two users editing the *same* square at the
same time.